# api/openapi/ops-platform.yaml
openapi: 3.0.3
info:
  title: Ops Platform API
  version: 0.1.0
servers:
  - url: /api/v1
paths:
...
  /overview:
    get:
      summary: Get cluster overview
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Cluster overview
  /namespaces:
    get:
      summary: List namespaces
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Namespace list
  /namespaces/{namespace}:
    get:
      summary: Get namespace resource associations and declared resource totals
      description: Returns current Namespace metadata, resource counts, active Pod requests/limits, Pods, top-level Workloads, Services, Ingresses, PVCs and UID-filtered Events. Resource totals are Kubernetes declarations, not metrics-server usage.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
      responses:
        "200":
          description: Namespace detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/NamespaceDetail"
  /nodes:
    get:
      summary: List nodes
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Node page
  /nodes/{name}:
    get:
      summary: Get node scheduling and declared resource detail
      description: Returns Node conditions, roles, addresses, taints, system information, capacity/allocatable, effective requests/limits of non-terminal scheduled Pods, associated Pods/Workloads and UID-filtered Events. It does not report realtime CPU or memory usage.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Name"
      responses:
        "200":
          description: Node detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/NodeDetail"
  /resources/yaml:
    get:
      summary: Export Kubernetes resource YAML
      security:
        - bearerAuth: []
      parameters:
        - name: kind
          in: query
          required: true
          schema:
            type: string
            enum:
              - node
              - namespace
              - pod
              - deployment
              - statefulset
              - daemonset
              - replicaset
              - job
              - cronjob
              - service
              - ingress
              - configmap
              - pvc
              - pv
              - storageclass
        - name: namespace
          in: query
          schema:
            type: string
        - name: name
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resource YAML
    put:
      summary: Update Kubernetes resource from YAML
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourceYAMLUpdateRequest"
      responses:
        "200":
          description: Resource YAML updated
  /namespaces/{namespace}/pods:
    get:
      summary: List pods
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Pod page
  /namespaces/{namespace}/pods/{pod}:
    get:
      summary: Get pod diagnostic detail
      description: Returns runtime state, current and previous container termination details, owner references, resolved controller chain, conditions and scheduling/network metadata.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Pod"
      responses:
        "200":
          description: Pod detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PodDetail"
    delete:
      summary: Delete pod
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Pod"
        - name: confirm
          in: query
          required: true
          schema:
            type: boolean
      responses:
        "200":
          description: Delete request submitted
...
  /namespaces/{namespace}/pods/{pod}/yaml:
    get:
      summary: Export pod YAML
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Pod"
      responses:
        "200":
          description: Pod YAML
...
  /namespaces/{namespace}/deployments:
    get:
      summary: List deployments
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Deployment page
  /namespaces/{namespace}/deployments/{name}:
    get:
      summary: Get deployment detail
      description: Returns rollout strategy and conditions plus Deployment-owned ReplicaSets, their Pods, and related Deployment/ReplicaSet/Pod events.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      responses:
        "200":
          description: Deployment detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/DeploymentDetail"
  /namespaces/{namespace}/deployments/{name}/yaml:
    get:
      summary: Export deployment YAML
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      responses:
        "200":
          description: Deployment YAML
...
  /namespaces/{namespace}/statefulsets:
    get:
      summary: List statefulsets
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: StatefulSet page
  /namespaces/{namespace}/statefulsets/{name}:
    get:
      summary: Get statefulset detail
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      responses:
        "200":
          description: StatefulSet detail
  /namespaces/{namespace}/statefulsets/{name}/yaml:
    get:
      summary: Export statefulset YAML
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      responses:
        "200":
          description: StatefulSet YAML
...
  /namespaces/{namespace}/daemonsets:
    get:
      summary: List daemonsets
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: DaemonSet page
  /namespaces/{namespace}/daemonsets/{name}:
    get:
      summary: Get daemonset detail
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      responses:
        "200":
          description: DaemonSet detail
  /namespaces/{namespace}/daemonsets/{name}/yaml:
    get:
      summary: Export daemonset YAML
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      responses:
        "200":
          description: DaemonSet YAML
...
  /namespaces/{namespace}/replicasets:
    get:
      summary: List replicasets
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: ReplicaSet page
  /namespaces/{namespace}/replicasets/{name}:
    get:
      summary: Get replicaset detail
      description: Returns the owning Deployment reference, selector, revision, conditions, directly owned Pods, and related ReplicaSet/Pod events. Direct ReplicaSet mutation remains disabled.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      responses:
        "200":
          description: ReplicaSet detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ReplicaSetDetail"
  /namespaces/{namespace}/jobs:
    get:
      summary: List jobs
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Job page
  /namespaces/{namespace}/jobs/{name}:
    get:
      summary: Get job detail
      description: Returns execution policy, status conditions, owning CronJob reference, directly owned Pods, and related Job/Pod events. Job deletion and rerun are intentionally not exposed; rerun requires creating a new Job with a new name.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      responses:
        "200":
          description: Job detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/JobDetail"
  /namespaces/{namespace}/cronjobs:
    get:
      summary: List cronjobs
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: CronJob page
  /namespaces/{namespace}/cronjobs/{name}:
    get:
      summary: Get cronjob detail
      description: Returns schedule policy, Job template policy, directly owned Jobs, descendant Pods, and related CronJob/Job/Pod events. Immediate run and deletion are intentionally not exposed.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      responses:
        "200":
          description: CronJob detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/CronJobDetail"
...
  /namespaces/{namespace}/services:
    get:
      summary: List services
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Service page
  /namespaces/{namespace}/services/{name}:
    get:
      summary: Get service detail
      description: Returns Service traffic policy and structured ports, EndpointSlice-backed endpoints with legacy Endpoints fallback, selector/targetRef-associated Pods, and related Service/endpoint/Pod events.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      responses:
        "200":
          description: Service detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ServiceDetail"
  /namespaces/{namespace}/ingresses:
    get:
      summary: List ingresses
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Ingress page
  /namespaces/{namespace}/ingresses/{name}:
    get:
      summary: Get ingress detail
      description: Returns default and rule backends, TLS metadata, de-duplicated backend Service details, EndpointSlice-backed endpoints with legacy Endpoints fallback, associated Pods, and related Ingress/Service/endpoint/Pod events. Missing backend Services remain visible for diagnosis.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      responses:
        "200":
          description: Ingress detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/IngressDetail"
  /namespaces/{namespace}/configmaps:
    get:
      summary: List configmaps
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: ConfigMap page
  /namespaces/{namespace}/persistentvolumeclaims:
    get:
      summary: List persistent volume claims
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: PVC page
  /namespaces/{namespace}/persistentvolumeclaims/{name}:
    get:
      summary: Get persistent volume claim storage associations
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      responses:
        "200":
          description: PVC detail with PV, Pod, workload, mount and Event associations
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PVCDetail"
  /persistentvolumes:
    get:
      summary: List persistent volumes
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: PV page
  /persistentvolumes/{name}:
    get:
      summary: Get persistent volume storage associations
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Name"
      responses:
        "200":
          description: PV detail with PVC, Pod, workload, mount and Event associations
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PVDetail"
  /storageclasses:
    get:
      summary: List storage classes
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: StorageClass page
  /namespaces/{namespace}/events:
    get:
      summary: List namespace events
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - name: involved_kind
          in: query
          schema:
            type: string
        - name: involved_name
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Event list
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Namespace:
      name: namespace
      in: path
      required: true
      schema:
        type: string
    Pod:
      name: pod
      in: path
      required: true
      schema:
        type: string
    Name:
      name: name
      in: path
      required: true
      schema:
        type: string
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        default: 20
  schemas:
    ResponseEnvelope:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          nullable: true
...
    NamespaceDetail:
      allOf:
        - $ref: "#/components/schemas/NamespaceSummary"
        - type: object
          properties:
            labels:
              type: object
              additionalProperties:
                type: string
            finalizers:
              type: array
              items:
                type: string
            conditions:
              type: array
              items:
                $ref: "#/components/schemas/PodCondition"
            counts:
              $ref: "#/components/schemas/NamespaceResourceCounts"
            allocated:
              $ref: "#/components/schemas/ResourceAllocation"
            pods:
              type: array
              items:
                $ref: "#/components/schemas/PodSummary"
            workloads:
              type: array
              items:
                $ref: "#/components/schemas/ResourceReference"
            services:
              type: array
              items:
                $ref: "#/components/schemas/ServiceSummary"
            ingresses:
              type: array
              items:
                $ref: "#/components/schemas/IngressSummary"
            persistent_volume_claims:
              type: array
              items:
                $ref: "#/components/schemas/PVCSummary"
            events:
              type: array
              items:
                $ref: "#/components/schemas/EventSummary"
...
    NodeDetail:
      allOf:
        - $ref: "#/components/schemas/NodeSummary"
        - type: object
          properties:
            roles:
              type: array
              items:
                type: string
            unschedulable:
              type: boolean
            pod_cidrs:
              type: array
              items:
                type: string
            labels:
              type: object
              additionalProperties:
                type: string
            addresses:
              type: array
              items:
                $ref: "#/components/schemas/NodeAddressDetail"
            taints:
              type: array
              items:
                $ref: "#/components/schemas/NodeTaintDetail"
            conditions:
              type: array
              items:
                $ref: "#/components/schemas/NodeConditionDetail"
            system_info:
              $ref: "#/components/schemas/NodeSystemInfo"
            capacity:
              $ref: "#/components/schemas/DeclaredResources"
            allocatable:
              $ref: "#/components/schemas/DeclaredResources"
            allocated:
              $ref: "#/components/schemas/ResourceAllocation"
            pods:
              type: array
              items:
                $ref: "#/components/schemas/PodSummary"
            workloads:
              type: array
              items:
                $ref: "#/components/schemas/ResourceReference"
            events:
              type: array
              items:
                $ref: "#/components/schemas/EventSummary"
...
    PodDetail:
      allOf:
        - $ref: "#/components/schemas/PodSummary"
        - type: object
          properties:
            init_containers:
              type: array
              items:
                $ref: "#/components/schemas/ContainerSummary"
            pod_ip:
              type: string
            host_ip:
              type: string
            qos_class:
              type: string
            service_account:
              type: string
            restart_policy:
              type: string
            start_time:
              type: string
              format: date-time
            labels:
              type: object
              additionalProperties:
                type: string
            annotations:
              type: object
              additionalProperties:
                type: string
            owner_refs:
              type: array
              items:
                $ref: "#/components/schemas/OwnerReference"
            controller_chain:
              type: array
              items:
                $ref: "#/components/schemas/OwnerReference"
            conditions:
              type: array
              items:
                $ref: "#/components/schemas/PodCondition"
...
    ResourceYAMLUpdateRequest:
      type: object
      required:
        - kind
        - namespace
        - name
        - yaml
      properties:
        kind:
          type: string
          enum:
            - deployment
            - statefulset
            - daemonset
            - job
            - cronjob
            - service
            - ingress
        namespace:
          type: string
        name:
          type: string
        yaml:
          type: string
...